Feat: mv search bar to permanent position, show only results in dropdown
