+++ src/app/components/common/Collapsible.tsx
@@ -4,11 +4,12 @@ import React, { useState } from 'react';
 type Props = {
     title: string;
     open?: boolean;
+    noPadding?: boolean;
     children: React.ReactNode;
 };
 
 const Collapsible: React.FC<Props> = (props) => {
-    const { title, open = false } = props;
+    const { title, open = false, noPadding } = props;
 
     const [isOpen, setIsOpen] = useState(open);
 
@@ -17,10 +18,14 @@ const Collapsible: React.FC<Props> = (props) => {
     };
 
     return (
-        <div className="bg-primary border-b text-black border-gray-300">
+        <div className="bg-primary  text-black ">
             <button
                 title={`${isOpen ? 'Hide' : 'Show'} ${title}`}
-                className="w-full flex justify-between items-center p-4 bg-primary font-bold text-lg hover:bg-primary-hover active:bg-primary-hover focus:bg-primary-hover"
+                className={`sticky top-0 w-full 
+                        flex justify-between items-center p-4 bg-primary-opaque 
+                        font-bold text-lg 
+                        border-t ${isOpen && 'border-b'} border-gray-300
+                        hover:bg-primary-opaque-hover focus:bg-primary-opaque-hover`}
                 onClick={toggleCollapse}
             >
                 {title}
@@ -37,7 +42,7 @@ const Collapsible: React.FC<Props> = (props) => {
                 data-testid="collapsible-content"
                 className={`overflow-hidden ${isOpen ? 'block' : 'hidden'}`}
             >
-                <div className="px-4">{props.children}</div>
+                {props.children}
             </div>
         </div>
     );
